requirements.txt
streamlit
python-binance
numpy
pandas
matplotlib
praw
yfinance
requests
textblob

